Remove Config::Irssi::Parser requirement
A bit of misunderstanding of how irssi worked internally lead me down
the path of reading the irssi configuration file to determine what
exactly was a highlight. This was overkill and not needed.

The issue was with how irssi let you know a certain thing was a highlight.
You could only see if the print_text signal was a highlight, and this
signal only passed in the text itself (not the other metadata).

To get around this, each signal we care about has a handler that grabs
the important data and stores it in a hash. When the print_text signal
happens, its handler pulls the data from the hash in to a string and
then sends it to the Python script via the shell.

This should be more resillient now, especially if someone has whacky
stuff in their configuration file (assuming Config::Irssi::Parser
would barf).